requirements.txt
opencv-python>=4.5.0
numpy>=1.19.5,<1.20.0
pyserial>=3.5
